Backend (API)/requirements.txt
Flask==3.0.3
flask-cors==6.0.1
mysql-connector-python==9.1.0
meta_ai_api==1.2.5